Use DbusMonitor instead of VeDbusItemImport
When using the (simpler) VeDbusItemImport without also listening
to nameownerchanged signals (like I was doing) the program won't
notice it when the dbus-service that provides the grid measurements
goes offline. It will just keep publishing the last received grid-
power and current.

DbusMonitor takes care of the NameOwnerChanged signal, and changes
the value to None.